Add elf and macho modules
